+++ src/main/java/avengers/lion/mission/dto/MissionPreReviewResponse.java
@@ -1,11 +1,12 @@
 package avengers.lion.mission.dto;
 
 
+import avengers.lion.global.base.PageMeta;
+
 import java.util.List;
 
 public record MissionPreReviewResponse(
         Long count,
         List<MissionReviewResponse> missionReviewResponse,
         PageMeta reviewPage) {
-    public record PageMeta(Boolean hasNext, Long nextId){};
 }
