+++ src/main/java/avengers/lion/review/dto/OverviewResponse.java
@@ -1,8 +1,8 @@
 package avengers.lion.review.dto;
 
+import avengers.lion.global.base.PageMeta;
 import io.swagger.v3.oas.annotations.media.Schema;
 
-import java.time.LocalDateTime;
 import java.util.List;
 
 @Schema(name = "OverviewResponse", description = "프리뷰: 작성 가능한리뷰, 작성한 리뷰 개수 + 각 탭 프리뷰 데이")
@@ -25,13 +25,5 @@ public record OverviewResponse(
         @Schema(description = "작성한 리뷰 프리뷰의 다음 페이지 메타")
         PageMeta writtenPage
 ) {
-    @Schema(name = "OverviewResponse.PageMeta", description = "다음 페이지 조회를 위한 커서 메타")
-    public record PageMeta(
-            @Schema(description = "다음 페이지 존재 여부", example = "true")
-            Boolean hasNext,
-            @Schema(description = "다음 페이지 시작 기준 시간 (LocalDateTime ISO-8601)", example = "2025-08-09T12:22:10")
-            LocalDateTime nextAt,
-            @Schema(description = "다음 페이지 시작 기준 ID", example = "5")
-            Long nextId
-    ) { }
+
 }
